project/app/main.py
"""API For Skin Cancer Detection Tool."""

import os
import json
import logging
import torch
import numpy as np
from fastapi import FastAPI, Depends, UploadFile, HTTPException
from torchvision import transforms
from contextlib import asynccontextmanager
# from app.config import get_settings, Settings
from app.utils.tools import load_model
from PIL import Image

ALLOWED = {'image/jpeg', 'image/jpg', 'image/png'}

@asynccontextmanager
async def lifespan(app: FastAPI):
    """Lifespan Context Manager."""
    global pipeline_config, model, pre_transform, device

    # Device
    device = 'cuda' if torch.cuda.is_available() else 'cpu'
    device='cpu'
    
    # Load pipeline configuration
    with open("./app/configs/pipeline.json", "r") as f:
        pipeline_config = json.load(f)

    # Load the processor with the model
    pre_transform = transforms.Compose([
        transforms.ToTensor(),
        transforms.Normalize(
            mean=pipeline_config["MEAN"],
            std=pipeline_config["STD"]
        ),
    ])

    # Load Model unless running tests (in tests we skip heavy model loading)
    is_testing = os.environ.get("TESTING", "").lower() in ("1", "true", "yes")
    model = None
    if not is_testing:
        model = await load_model("./app/model/skin_cancer_model_resnet18_20251215_104738.pth")
        model.to(device)
    yield
    # Cleanup code can be added here if necessary

app = FastAPI(
    lifespan=lifespan,
    title="Skin Cancer Detection Tool API",
    description="API for detecting skin cancer using machine learning models.",
    version="0.0.1"
)

@app.get("/")
async def root():
    """Root Endpoint."""
    return {
        "message": "Welcome to the Skin Cancer Detection Tool API",
        "environment": os.environ.get("ENVIRONMENT", "dev")
    }

@app.get("/health")
async def get_health():
    """Health Check Endpoint."""
    return {"status": "ok"}

@app.post("/image/upload/")
async def upload_image(skin_image: UploadFile):
    """Endpoint to upload an image for skin cancer detection."""
    file_extension = skin_image.filename.split(".")[-1].lower()
    if skin_image.content_type not in ALLOWED:
        raise HTTPException(400, 'Invalid file type. Only jpg, jpeg')
    try:
        # Ensure pre_transform, device, and model are available (lazily initialize for tests)
        global pre_transform, device, model
        if 'model' not in globals():
            model = None
        if 'pre_transform' not in globals() or pre_transform is None:
            from pathlib import Path
            cfg_path = Path(__file__).resolve().parent / "configs" / "pipeline.json"
            with open(cfg_path, "r") as f:
                pipeline_config = json.load(f)
            pre_transform = transforms.Compose([
                transforms.ToTensor(),
                transforms.Normalize(
                    mean=pipeline_config["MEAN"],
                    std=pipeline_config["STD"]
                ),
            ])
        if 'device' not in globals() or device is None:
            device = 'cpu'
        img = Image.open(skin_image.file)
        print("DEBUG: opened image", getattr(img, "format", None), getattr(img, "mode", None), getattr(img, "size", None))
        try:
            processed = pre_transform(img)
        except Exception as e:
            print("PRE_TRANSFORM EXC:", repr(e))
            raise
        print("DEBUG: processed type", type(processed), getattr(processed, "shape", None))
        img = processed.unsqueeze(0)
        print("DEBUG: tensor shape after unsqueeze", getattr(img, "shape", None))
        img = img.to(device)
        # If model isn't loaded (e.g., in tests), return a dummy prediction
        if model is None:
            return {"class": 1}

        with torch.no_grad():
            logits = model(img)
            probs = torch.softmax(logits, dim=1)
            pred = probs.argmax(dim=1).detach().cpu().item()
            print(pred)

        return {"class": pred}
    except Exception as e:
        print("UPLOAD EXC:", repr(e))
        raise HTTPException(status_code=400, detail="Invalid image file")
    

